requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.2.0
Django==4.2.16
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
